backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
SQLAlchemy==2.0.16
sentence-transformers=2.2.2
python-dotenv==1.0.0apturl==0.5.2
